+++ modules/Scheduling/Courses.php
@@ -69,7 +69,7 @@ if ( isset( $_REQUEST['course_modfunc'] )
 
 	echo '<form name="search" action="Modules.php?modname=' . $_REQUEST['modname'] .
 		'&modfunc=' . $_REQUEST['modfunc'] . '&course_modfunc=search&last_year=' .
-		$_REQUEST['last_year'] . '" method="GET">';
+		$_REQUEST['last_year'] . '" method="POST">'; // Fix Search: Use POST for Public Pages plugin compatibility.
 
 	echo '<table><tr><td><input type="text" name="search_term" value="' .
 		issetVal( $_REQUEST['search_term'], '' ) . '" required autofocus /></td>