+++ CHANGES.md
@@ -23,6 +23,7 @@ Changes in 5.9
 - Moodle circumvent bug: no response or error but User created in Moodle/functions.php, User.php & Student.php
 - HTML add .br-after span to Course Periods + use DBGetOne() in Side.php
 - jQuery already loaded on first-login page, remove it in FirstLogin.fnc.php
+- CSS WPadmin pointer cursor for radio & checkbox input in stylesheet.css
 
 Changes in 5.8.1
 ----------------