+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 5.8.1
 - Fix Next Year widget: Handle "Retain" case: value is '0' in Widgets.fnc.php
 - Fix Search: Use POST for Public Pages plugin compatibility in Courses.php
 - Fix MarkDown preview + bold in warehouse.js & Inputs.php
+- Fix saving REMOVE_ACCESS_USERNAME_PREFIX_ADD on Grant in RemoveAccess.php
 
 Changes in 5.8
 --------------